dom-inspector/element-picker: Ensure frame background is transparent.
This fixes the element picker on dark color-scheme websites on
Firefox Nightly and Chrome Canary at least, see:

  * https://bugzilla.mozilla.org/show_bug.cgi?id=1771047
  * https://github.com/w3c/csswg-drafts/issues/4772

This also removes the existing (not quite correct, was relying on a
Chromium bug) fix for Chromium browsers.

The dom inspector scriptlet had the same issue (so I fixed it as well).